move test separately and fix go test comand
