+++ .github/workflows/kind_integration.yml
@@ -96,7 +96,8 @@ jobs:
 
     - name: Run CLI Integration tests
       run: |
-        go test --failfast --mod=readonly ".\test\cli_install_static_test.go" --linkerd=$HOME\linkerd.exe --integration-tests TestCliInstall
+        # Run go test
+        go test --failfast --mod=readonly ".\test\static\cli_install_static_test.go" --linkerd=$HOME\linkerd.exe --linkerd-namespace=l5d  TestCliInstall -v  
   kind_integration_tests:
     strategy:
       matrix:
